% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{sn_fetch_cistarget_resources}
\alias{sn_fetch_cistarget_resources}
\title{Fetch Cistarget resources}
\usage{
sn_fetch_cistarget_resources(
  species = c("human", "mouse"),
  outdir = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{species}{The species to use for the analysis. Default: \code{"human"}.}

\item{outdir}{The directory to save the results. Default: \code{NULL}.}

\item{overwrite}{Whether to overwrite the results. Default: \code{FALSE}.}
}
\value{
A list of local paths to the downloaded files.
}
\description{
Fetch Cistarget resources
}
\examples{
\dontrun{
sn_fetch_cistarget_resources(species = "human")
}

}
