% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{sn_run_pyscenic}
\alias{sn_run_pyscenic}
\title{Run PySCENIC on a Seurat Object}
\usage{
sn_run_pyscenic(
  x,
  assay = "RNA",
  layer = "counts",
  outdir = "./pyscenic",
  species = "human",
  features = NULL,
  tfs_fname = NULL,
  database_fname = NULL,
  annotations_fname = NULL,
  mode = "custom_multiprocessing",
  ncores = 8,
  seed = 717,
  overwrite = FALSE
)
}
\arguments{
\item{x}{A Seurat object.}

\item{assay}{The assay to use for the analysis. Default: \code{"RNA"}.}

\item{layer}{The layer to use for the analysis. Default: \code{"counts"}.}

\item{outdir}{The directory to save the results. Default: \code{"./pyscenic"}.}

\item{species}{The species to use for the analysis. Default: \code{"human"}.}

\item{features}{The features to use for the analysis. Default: \code{NULL}.}

\item{tfs_fname}{The path to the TF list file. Default: \code{NULL}.}

\item{database_fname}{The path to the database file. Default: \code{NULL}.}

\item{annotations_fname}{The path to the annotations file. Default: \code{NULL}.}

\item{mode}{The mode to use for the analysis. Default: \code{"custom_multiprocessing"}.}

\item{ncores}{The number of cores to use for the analysis. Default: \code{8}.}

\item{seed}{The seed to use for the analysis. Default: \code{717}.}

\item{overwrite}{Whether to overwrite the results. Default: \code{FALSE}.}
}
\value{
A list of results. The results are saved in the \code{outdir} directory.
}
\description{
Run PySCENIC on a Seurat Object
}
\examples{
\dontrun{
sn_run_pyscenic(seurat_obj, species = "human")
}

}
